+++ flowserv/model/parameter/numeric.py
@@ -337,3 +337,18 @@ def is_int(para: ParameterBase) -> bool:
     bool
     """
     return para.type_id == PARA_INT
+
+
+def is_numeric(para: ParameterBase) -> bool:
+    """Test if the given parameter is of type PARA_FLOAT or PARA_INT.
+
+    Parameters
+    ----------
+    para: flowserv.model.parameter.base.ParameterBase
+        Template parameter definition.
+
+    Returns
+    -------
+    bool
+    """
+    return para.type_id in NUMERIC_TYPES