+++ changelog.md
@@ -79,3 +79,4 @@
 ### 0.5.1 - 2020-10-02
 
 * Create directory for SQLite database files (\#68).
+* Type tests for enum and numeric parameters.